requirements.txt
# Core automation dependencies
playwright>=1.40.0,<2.0.0
selenium>=4.15.0,<5.0.0
httpx>=0.25.0,<1.0.0
python-decouple>=3.8,<4.0.0
structlog>=23.1.0,<24.0.0

# Testing dependencies
pytest>=7.4.0,<8.0.0
pytest-playwright>=0.4.0,<1.0.0
pytest-mock>=3.11.0,<4.0.0
pytest-asyncio>=0.21.0,<1.0.0

# Development tools
black>=23.7.0,<24.0.0
isort>=5.12.0,<6.0.0
flake8>=6.0.0,<7.0.0
mypy>=1.5.0,<2.0.0
